refactoring debug notice

+++ src/debug-notice/init-debug-notice-event-listeners.js
@@ -2,9 +2,11 @@ import temporaryGlobals from "../temporary-globals/temporary-globals";
 
 export default function initDebugNoticeEventListeners () {
   if (temporaryGlobals.showDebugNotice) {
-    temporaryGlobals.DEBUG_RESET_ELEM.addEventListener("click", e => {
-      window.localStorage.clear(); 
-      window.location.reload();
+    document.addEventListener("click", event => {
+      if (event.closest(temporaryGlobals.SELECTORS.DEBUG_ELEM_RESET_SELECTOR)) {
+        window.localStorage.clear(); 
+        window.location.reload();
+      }
     });
   }
 }
